requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
cssselect2==0.7.0
Django==5.1.5
django-environ==0.12.0
djangorestframework==3.15.2
fonttools==4.55.4
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
weasyprint==63.1
webencodings==0.5.1
zopfli==0.2.3.post1
